Split general APR pool from evaluation heap and tracker.
Wrap-APR is now afw_pool_create(); the previous prefix/first-fit
implementation is afw_pool_heap_create / heap_tracker. Job pools set
managed_p to self. compiled_value evaluate creates a heap, clones the
result, and releases the heap.
