+++ src/afw/generate/interfaces/afw_interface.xml
@@ -5700,12 +5700,20 @@ AFW_DEFINE(const afw_request_handler_t *)
 
     <description>
       Hierarchical memory pool: fast allocate with bulk free on destroy or
-      release. Most AFW values, objects, and scopes allocate from pools;
-      subpools scope lifetime under a parent. Prefer afw_pool_create* /
-      afw_pool_calloc helpers and call macros over managing APR pools by
-      hand. Do not free individual allocations unless the pool supports it
-      (free_memory path). See group afw_pool.
+      release. Most AFW values, objects, and scopes allocate from pools.
+      Prefer afw_pool_create* / afw_pool_calloc helpers and call macros over
+      managing APR pools by hand. Do not free individual allocations unless
+      the pool supports it (free_memory path). See group afw_pool.
     </description>
+
+    <variable name="managed_p" type="const afw_pool_t *">
+      <description>
+        Pool to use for managed object/array instances and other managed
+        allocations made while this pool is the current p. Never NULL after
+        create. Heap and job pools (xctx, factory, thread, env) set this to
+        self; other pools inherit the parent.
+      </description>
+    </variable>
 
     <method name="release">
