+++ src/afw/object/afw_object_memory.c
@@ -38,7 +38,7 @@ afw_object_create_with_options(
     /* If managed, create subpool for object. */
     if (options == AFW_OBJECT_MEMORY_OPTION_managed)
     {
-        p = afw_pool_create_subpool(p, xctx);
+        p = afw_pool_create(p->managed_p, xctx);
     }
 
     /* Allocate memory for self and initialize. */
